2_REST_Application/requirements.txt
Flask
pydantic
alembic
black
isort
pytest-cov